feat: 이미지정리
