+++ public/assets/emotionChart/emotion.ts
@@ -1,0 +1,15 @@
+export const BG_GRAY_EMOTION = {
+  HAPPY: '/assets/emotionChart/emoji-smile.svg',
+  MOVED: '/assets/emotionChart/emoji-heart.svg',
+  WORRIED: '/assets/emotionChart/emoji-worry.svg',
+  SAD: '/assets/emotionChart/emoji-sad.svg',
+  ANGRY: '/assets/emotionChart/emoji-mad.svg',
+};
+
+export const UNCLICKED_EMOTION = {
+  HAPPY: '/assets/emotionChart/ic-unclicked-happy.svg',
+  MOVED: '/assets/emotionChart/ic-unclicked-love.svg',
+  WORRIED: '/assets/emotionChart/ic-unclicked-worry.svg',
+  SAD: '/assets/emotionChart/ic-unclicked-sad.svg',
+  ANGRY: '/assets/emotionChart/ic-unclicked-mad.svg',
+};
