+++ src/components/ChartContainer.tsx
@@ -1,4 +1,4 @@
-import { IMG_EMOTION } from '@/public/assets/emotionChart';
+import IMG_EMOTION from '@/public/assets/emotionChart';
 import { EmotionData } from '@/src/types/emotion';
 
 import instance from '@/src/app/api/axios';
